Remove lexical binding file local variables
Lexical binding was being set to t in the file local vars at the end of
each file. The Elisp manual states that this value is special and must
be set on the first line.

Since lexical binding is not in fact being used, setting the var on the
prop line could introduce bugs. Therefore it is safer to simply remove
the variable.